Adjust the order of initialization of the structure database and the reaction objects when creating a marley::Generator using a marley::JSONConfig object.
WARNING: Initializing the reactions before the structure database currently
prevents the reactions from seeing the discrete level data added later!
This issue is fixed by this commit for generators created by
marley::JSONConfig (and it never existed for those made with
marley::ConfigFile), but it is now a known bug in the MARLEY LArSoft module
(MARLEYGen_module.cc in LArSoft v06_18_00 and older).

This commit fixes the problem for marley::JSONConfig, but
marley::NuclearReaction should be refactored in the future (the relevant
code is from MARLEY's early days) to ensure that it always sees all
available structure data.